buttons in pause menu are now movable

+++ Core/Source/Scene_Inventory.cpp
@@ -115,7 +115,7 @@ void Scene_Inventory::spawnItem(std::string sprite)
     auto entityID = m_ECS.addEntity();
     m_item = entityID;
     std::cout << entityID << std::endl;
-    Vec2 pos = {(int)(entityID-1)%(int)m_inventorySize.x, (int)(entityID-1)/(int)m_inventorySize.x};
+    Vec2 pos = {(float)((int)(entityID-1)%(int)m_inventorySize.x), (float)((int)(entityID-1)/(int)m_inventorySize.x)};
     m_ECS.addComponent<CTransform>(entityID, pos, Vec2{0,0}, Vec2{4, 4}, 0.0f, 0.0f, true);
     m_ECS.addComponent<CBoundingBox>(entityID, Vec2 {32, 32});
